NFX Version 1.0.2

+++ PKGBUILD
@@ -1,6 +1,6 @@
 pkgname=nfx-bin
-pkgver=1.0.1
-pkgrel=2
+pkgver=1.0.2
+pkgrel=1
 pkgdesc="Cross-platform package manager for the Pheonix Studios ecosystem"
 
 arch=('x86_64')
@@ -14,13 +14,15 @@ depends=('glibc')
 provides=('nfx')
 conflicts=('nfx')
 
+groups=('pheonix-ecosystem')
+
 options=('!strip')
 
 source=(
-    "https://pheonix-studios-git.github.io/PPI/data/NFX/nfx_zip/NFX-v${pkgver}.zip"
+    "NFX-v${pkgver}.zip::https://pheonix-studios-git.github.io/PPI/data/NFX/nfx_zip/NFX-v${pkgver}.zip"
 )
 
-sha256sums=('c343097feeb001e554eec6c44f58c2353a551f90b649ee8e23a7f7b8f9dd6952')
+sha256sums=('e3085c52e346780c02dcfcd79f5d1b93525e761b570126f9adce56a7582f05dd')
 
 prepare() {
     cd "$srcdir"
@@ -36,11 +38,15 @@ package() {
         "$srcdir/bin/linux/x86_64/dist/nfx" \
         "$pkgdir/usr/bin/nfx"
 
-    install -Dm644 \
-        "$srcdir/LICENSE" \
-        "$pkgdir/usr/share/licenses/$pkgname/LICENSE"
+    if [[ -f "$srcdir/LICENSE" ]]; then
+        install -Dm644 \
+            "$srcdir/LICENSE" \
+            "$pkgdir/usr/share/licenses/$pkgname/LICENSE"
+    fi
 
-    install -Dm644 \
-        "$srcdir/README.md" \
-        "$pkgdir/usr/share/doc/$pkgname/README.md"
+    if [[ -f "$srcdir/README.md" ]]; then
+        install -Dm644 \
+            "$srcdir/README.md" \
+            "$pkgdir/usr/share/doc/$pkgname/README.md"
+    fi
 }
